+++ src/domain/body/fallingBody/FallingBody.java
@@ -5,10 +5,10 @@ import domain.body.Body;
 public abstract class FallingBody extends Body {
 
 
-    public FallingBody(double x_coordinates,
-                       double y_coordinates,
-                       double length,
-                       double width) {
+    public FallingBody(int x_coordinates,
+                       int y_coordinates,
+                       int length,
+                       int width) {
         super(x_coordinates, y_coordinates, length, width);
     }
     public abstract void fall();
